[integration-tests] correctly reset static metrics between tests
The maven-surefire upgrade to 3.0.0-M5 helped debug why two of the tests were
silently failing - metrics were not being reset correctly for the first two
test classes.  Before M5 the tests were not running (or failing) at all in CI
or via the maven command line, and no output was provided.

This adds in another hack to make sure that the static metrics are reset
correctly.  It all works but...